refactor: unify HT and TTC calculations
